Moved Database, changed methods
